tests/justhtml-tests/document_shell_recovery.dat

#data
</template>
#errors
(1,1): expected-doctype-but-got-end-tag
(1,11): unexpected-end-tag-before-html

#document
| <html>
|   <head>
|   <body>

#data
<body><input type="hidden">
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <input>
|       type="hidden"

#data
<html><body><html attr="test">
#errors
(1,1): expected-doctype-but-got-start-tag
(1,30): unexpected-start-tag

#document
| <html>
|   attr="test"
|   <head>
|   <body>

#data
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
#errors
(1,100): unknown-doctype

#document
| <!DOCTYPE html "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|   <body>

#data
<head></head></template>text
#errors
(1,1): expected-doctype-but-got-start-tag
(1,24): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<meta><!--x-->
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <meta>
|     <!-- x -->
|   <body>

#data
<!DOCTYPE 
#errors
(1,10): eof-in-tag
(1,10): unknown-doctype

#document
| <!DOCTYPE  >
| <html>
|   <head>
|   <body>

#data
<!doctype html>x<head><meta>
#errors

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "x"
|     <meta>

#data
<!DOCTYPE html foo><p><table>
#errors
(1,19): unknown-doctype

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <table>


#data
<div a="unterminated
#errors
#document
| <html>
|   <head>
|   <body>

#data
<DIV A=1 a=2>x</DIV>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       a="1"
|       "x"

#data
<! first><?second
#errors
(1,1): expected-doctype-but-got-chars
#document
| <!--  first -->
| <html>
|   <head>
|   <body>

#data
<div \x00name=value>
#errors
(1,6): unexpected-null-character
(1,17): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <div>
|       �name="value"

#data
<source\x00 x=y>
#errors
(1,8): unexpected-null-character
(1,13): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <source�>
|       x="y"

#data
<html 1name='\x00'>
#errors
(1,14): unexpected-null-character
(1,16): expected-doctype-but-got-start-tag
#document
| <html>
|   1name="�"
|   <head>
|   <body>

#data
<source\x00 '='<script>alert(1)</script>'>
#errors
(1,8): unexpected-null-character
(1,20): expected-doctype-but-got-start-tag
(1,29): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <source�>
|       '="<script>alert(1)</script>"

#data
</html><script>
#errors
(1,1): expected-doctype-but-got-chars
(1,1): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <script>

#data
</head></html><y>
#errors
(1,1): expected-doctype-but-got-chars
(1,1): unexpected-end-tag
(1,8): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <y>

#data
</html><</\x20
#errors
(1,1): expected-doctype-but-got-chars
(1,1): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     "<"
|     <!--   -->

#data
<i></body></f><
#errors
(1,3): expected-doctype-but-got-start-tag
(1,4): unexpected-end-tag
(1,11): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <i>
|       "<"

#data
</>&#9
#errors
(1,1): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>

#data
<!>&#xD
#errors
(1,1): expected-doctype-but-got-chars
#document
| <!--  -->
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE>&#xD
#errors
(1,10): unknown-doctype
#document
| <!DOCTYPE >
| <html>
|   <head>
|   <body>

#data
</e>&#9
#errors
(1,1): expected-doctype-but-got-chars
(1,1): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>

#data
</head></html>2</t></-
#errors
(1,1): expected-doctype-but-got-chars
(1,1): unexpected-end-tag
(1,8): unexpected-end-tag
(1,16): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     "2"
|     <!-- - -->

#data
<body id='a'><div><body id='b'>x</div>
#errors
(1,13): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     id="a"
|     <div>
|       "x"

#data
<div><li>x<body class='late'>y
#errors
(1,5): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     class="late"
|     <div>
|       <li>
|         "xy"

#data
<div>x<head><title>ignored</title></head>y
#errors
(1,5): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <title>
|         "ignored"
|       "y"

#data
<!DOCTYPE html><body><command>A
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <command>
|       "A"

#data
<head></head>\x0a
#errors
#document
| <html>
|   <head>
|   "
"
|   <body>

#data
<!doctype html><div></span\x0a>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>

#data
İ<!DOCTYPE html><p>x</p>
#errors
#document
| <html>
|   <head>
|   <body>
|     "İ"
|     <p>
|       "x"
